refactor(context): move application context next to the application component
