+++ app/scripts/modules/core/src/application/ApplicationComponent.tsx
@@ -2,6 +2,7 @@ import { ApplicationNavigation } from './nav/ApplicationNavigation';
 import React from 'react';
 import { UIView } from '@uirouter/react';
 
+import { ApplicationContextProvider } from './ApplicationContext';
 import { Application } from './application.model';
 import { RecentHistoryService } from 'core/history';
 import { DebugWindow } from 'core/utils/consoleDebug';
@@ -68,9 +69,11 @@ export class ApplicationComponent extends React.Component<IApplicationComponentP
             </p>
           </div>
         )}
-        <div className="container scrollable-columns">
-          <UIView className="secondary-panel" name="insight" />
-        </div>
+        <ApplicationContextProvider app={app}>
+          <div className="container scrollable-columns">
+            <UIView className="secondary-panel" name="insight" />
+          </div>
+        </ApplicationContextProvider>
       </div>
     );
   }
